+++ chromium/recorder_bridge/recorder_switches.h
@@ -5,6 +5,10 @@ namespace a11y_recorder {
 
 inline constexpr char kBootstrapSwitch[] = "a11y-recorder-bootstrap";
 inline constexpr char kBootstrapFromStandardInput[] = "stdin";
+inline constexpr char kChildBootstrapHandleSwitch[] =
+    "a11y-recorder-bootstrap-handle";
+inline constexpr char kChildProcessIdSwitch[] =
+    "a11y-recorder-child-process-id";
 
 }  // namespace a11y_recorder